Fix issue 8 (negative height)
Addresses issue 8, where bitmaps may be represented "top-down" rather than the standard "bottom-up".  The height in the header is represented as a signed integer, and if that number is negative, then the image data is "top-down".  In that case, a flag is set in the header (this really is header metadata, even though it is represented strangely).  In the iterator that is used for raw pixel data, this flag is used to determine whether the data is scanned starting from the bottom, rather than the top.