Add support for RLE encoded bitmaps. (#41)
* Add experimental support for RLE8 encoded bitmaps.

* Update CHANGELOG and README.

* Now you can display RLE8 bitmaps.

Also adds tests and benchmarks.

* Add RLE4 support.

The count is of nibbles, but the padding only applies when there's an odd number of bytes. And the sequences alternate between nibbles. But otherwise it's similar to RLE8.

* Increase MSRV to 1.71

* Applying review comments.

* Faster RLE rendering.

We can't render a whole frame as a contiguous block, we but we can render each line contiguously - we just have to start at the bottom line and work up.

---------

Co-authored-by: Ralf Fuest <mail@rfuest.de>