chore: disable warning for `header_type` of `DibHeader`
`header_type` is never read and `DibHeader` is private, so the compiler
emits a warning. I added `#[expect(unused)]` so if it gets used at some
point, compiler will ask you to remove it. `#[allow(dead_code)]` is
worse because it will itself become a dead code if `header_type` gets
used.