Make raw color iterators public (#51)
* feat: expose more structs to the user

* chore: changelog entry

* chore: disable warning for `header_type` of `DibHeader`

`header_type` is never read and `DibHeader` is private, so the compiler
emits a warning. I added `#[expect(unused)]` so if it gets used at some
point, compiler will ask you to remove it. `#[allow(dead_code)]` is
worse because it will itself become a dead code if `header_type` gets
used.

* refactor: group logic for rle pixel coordinates

In preparation of remove that logic from them and moving it up.

* feat: `RawBmp::colors`

Additional work that was required by `RawBmp::colors`:
- Making `DynamicRawColors` public.
- Making `DynamicRawColors` implement iterator.
- Refactoring logic.
- Rename `Rle{4,8}Pixels` to `Rle{4,8}Colors`
- Due to removal of position handling in `Rle{4,8}Pixels`, they can no
  longer assert the the new line corresponds with the width. I left the
  assertion functionally the same, but with slightly different approach.
  Personally, I don't think it is necessary that that assert there at
  all - like, either way the image would be broken and developer will
  have to fix the `bmp` file, that assertion isn't helpful at all.

* chore: changelog

* fix: remove unused `header_type` from private `DibHeader`

* fix: review comments

* fix: incorrect test, image has bottom-up order but expectation is top-down order